+++ src/main/kotlin/restApi/Models.kt
@@ -4,7 +4,7 @@ data class InvalidGroupChatPicUpdate(val reason: Reason) {
     enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE }
 }
 
-data class InvalidAudioMessage(val reason: Reason) {
+data class InvalidMediaMessage(val reason: Reason) {
     enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE, INVALID_CONTEXT_MESSAGE }
 }