+++ src/test/kotlin/db/DbUtil.kt
@@ -16,6 +16,7 @@ private val tables: List<Table> = listOf(
     TextMessages,
     PicMessages,
     AudioMessages,
+    VideoMessages,
     PollVotes,
     PollOptions,
     PollMessages,